kl divergence sparsity constraint

+++ mars_triplepalooza.py
@@ -157,8 +157,8 @@ def main(margin,output_size,learning_rate,is_overfitting):
 
     dstr1 = tf.contrib.distributions.Gamma(.5,.5)
     wmean, wvar = tf.nn.moments(weights, axes=[0,1])
-    alpha = (wmean/tf.sqrt(wvar)).astype('float32')
-    beta = (a2/wmean).astype('float32')
+    alpha = (wmean/tf.sqrt(wvar))
+    beta = (a2/wmean)
     dstr2 = tf.contrib.distributions.Gamma(alpha,best)
 
     loss2 = lmbd * tf.contrib.distributions.kl_divergence(dstr1, dstr2)
